Clarify distinction between ProposalEquality and Artifact Equality
