+++ contracts/interfaces/IArtifactRegistry.sol
@@ -6,14 +6,14 @@ import "@openzeppelin/contracts/token/ERC721/IERC721.sol";
 contract IArtifactRegistry is IERC721 {
   struct Artifact {
     address artist;
-    string _title;
-    string _artistName;
-    string _artistNationality;
-    string _artistBirthYear; // TODO(mm5917): better data type
-    string _created; // TODO(mm5917): better data type
-    string _medium;
-    string _size;
-    string _metaUri;
+    string title;
+    string artistName;
+    string artistNationality;
+    string artistBirthYear; // TODO(mm5917): better data type
+    string created; // TODO(mm5917): better data type
+    string medium;
+    string size;
+    string metaUri;
   }
 
   function mint(address who, Artifact memory _artifact) public returns (uint256);
